src/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.2.1
ultralytics==8.3.202
pillow==11.3.0
python-jose==3.3.0
#aiofiles==24.1.0
requests
gtts

